+++ ICD.Connect.Panels.Crestron/Devices/Hr/IHrRemoteAdapter.cs
@@ -1,0 +1,8 @@
+﻿using ICD.Connect.Panels.HardButtons;
+
+namespace ICD.Connect.Panels.Crestron.Devices.Hr
+{
+	public interface IHrRemoteAdapter : IHardButtonDevice
+	{
+	}
+}