+++ ICD.Connect.Panels.Crestron/Devices/Hr/IHrRemoteAdapterSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Misc.InfinetEx;
+
+namespace ICD.Connect.Panels.Crestron.Devices.Hr
+{
+	public interface IHrRemoteAdapterSettings : IInfinetExDeviceSettings
+	{
+		 
+	}
+}