+++ ICD.Connect.Panels.CrestronPro/Remotes/HrRemotes/Abstracts/AbstractHrRemoteAdapterSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Misc.CrestronPro.InfinetEx;
+using ICD.Connect.Panels.Crestron.Devices.Hr;
+
+namespace ICD.Connect.Panels.CrestronPro.Remotes.HrRemotes.Abstracts
+{
+	public abstract class AbstractHrRemoteAdapterSettings : AbstractInfinetExAdapterSettings, IHrRemoteAdapterSettings
+	{
+	}
+}