+++ ICD.Connect.Panels.CrestronPro/Remotes/HrRemotes/Hr310/Hr310AdapterSettings.cs
@@ -1,0 +1,10 @@
+﻿using ICD.Connect.Panels.CrestronPro.Remotes.HrRemotes.Abstracts;
+using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Panels.CrestronPro.Remotes.HrRemotes.Hr310
+{
+	[KrangSettings("Hr310Adapter", typeof(Hr310Adapter))]
+	public sealed class Hr310AdapterSettings : AbstractHrRemoteAdapterSettings
+	{
+	}
+}